feat: Add embark-status-plugin
Add the embark-status-plugin to the dapp.

The issue with making this in to a PR is that the user’s machine IP and device IP are not static and need to be configured correctly.

In addition, the IPFS config changes also require the machine’s IP to be used as the API endpoint and IPFS gateway, which may not work with other dapps.